Extracted setting device method
