+++ exasol_transformers_extension/udfs/question_answering_udf.py
@@ -1,8 +1,9 @@
 import torch
 import pandas as pd
 import transformers
-from typing import Tuple, List, Optional
+from typing import Tuple, List
 from exasol_transformers_extension.udfs import bucketfs_operations
+from exasol_transformers_extension.utils import device_management
 
 
 class QuestionAnswering:
@@ -26,7 +27,7 @@ class QuestionAnswering:
 
     def run(self, ctx):
         device_id = ctx.get_dataframe(1).iloc[0]['device_id']
-        self._set_device(device_id)
+        self.device = device_management.get_torch_device(device_id)
         ctx.reset()
 
         while True:
@@ -38,16 +39,6 @@ class QuestionAnswering:
             ctx.emit(result_df)
 
         self.clear_device_memory()
-
-    def _set_device(self, device_id: Optional[int]) -> None:
-        """
-        Set device to push models
-
-        :param device_id: Either the id of cuda device or None implying CPU
-        """
-        device_name = f"cuda:{device_id}" \
-            if torch.cuda.is_available() and device_id is not None else "cpu"
-        self.device = torch.device(device_name)
 
     def get_batched_predictions(self, batch_df: pd.DataFrame) -> pd.DataFrame:
         """
